adding collections class, code before changing client to be more like a builder
