+++ src/main/java/org/example/Collections.java
@@ -1,0 +1,8 @@
+package org.example;
+
+/**
+ * Class for all the Collections Operations
+ */
+public class Collections {
+
+}
